+++ src/mbed_cloud/sdk/sdk.py
@@ -20,7 +20,6 @@ from mbed_cloud.client.client import Client
 from mbed_cloud.foundation.entities.entity_factory import EntityFactory
 
 global_sdk = None
-has_warned = None
 
 
 class SDK(object):
@@ -42,7 +41,6 @@ class SDK(object):
             configuration
         :type host: str
         """
-        beta_warning(self.__class__)
         check_openssl_version()
 
         # create a new config based on those we received
@@ -94,19 +92,6 @@ def get_or_create_global_sdk_instance():
     if global_sdk is None:
         global_sdk = SDK()
     return global_sdk
-
-
-def beta_warning(header, category=FutureWarning):
-    """Utility function to generate a beta warning."""
-    global has_warned
-    if not has_warned:
-        warnings.warn(
-            "[Beta] this section of the SDK is at a `beta` release level and is subject to change without notice: %s"
-            % header,
-            category=category,
-            stacklevel=3,  # make the trace log from two levels above this `warnings.warn` line
-        )
-        has_warned = True
 
 
 def check_openssl_version():
